requirements.txt
geopandas==0.13.2
fiona==1.9.6
hydra-core==1.3.2
ipython==8.12.0
matplotlib==3.5.3
mmcv==1.3.12
openmim
natsort==8.4.0
numpy==1.24.4
open3d==0.18.0
opencv_python==4.9.0.80
pandas==2.0.3
PDAL==2.3.0
laspy[lazrs,laszip]
Pillow
pqdict==1.4.0
pytest==8.3.5
pytorch_lightning==2.1.3
rospkg==1.5.0
rospy==1.16.0
scipy
segmentation_models_pytorch==0.3.3
setuptools==68.2.2
shapely==2.0.4
scikit-image
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
yacs==0.1.8
